ai-chatbot-be/requirements.txt
# FastAPI & Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.17
python-dotenv==1.0.1
websockets==14.1
aiofiles==24.1.0
sse-starlette>=3.0.3
httpx-sse>=0.4.3
anyio==4.7.0

# Database & Caching
supabase==2.10.0
psycopg2-binary
redis>=5.0.0

# LangChain Core
langchain==0.3.11
langchain-core==0.3.28
langchain-openai==0.2.10
langchain-community==0.3.11
langchain-huggingface==0.1.2
langchain-ollama==0.2.2
langchain-aws==0.2.6
langchain-anthropic>=0.2.0
langchain-mistralai>=0.2.0
langsmith

# LangGraph (Stateful RAG Pipeline)
langgraph>=0.2.0
langgraph-checkpoint>=2.0.0
langgraph-checkpoint-postgres>=2.0.0

# Embeddings & Vector Search
sentence-transformers>=3.0.0
faiss-cpu
tiktoken>=0.5.0

# Document Processing
pypdf2==3.0.1
python-docx==1.1.2
unstructured>=0.10.0

# Entity Extraction (NLP)
spacy>=3.7.0

# Date/Time Parsing
python-dateutil>=2.8.0

# Pydantic & Validation
pydantic==2.10.3
pydantic-settings==2.7.0

# Authentication & Security
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.5.0

# Email
fastapi-mail==1.4.1
jinja2==3.1.4

# Google Services
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.158.0



# AWS
boto3==1.35.64
botocore==1.35.64

# Development
black==23.12.1